fix: replace createGenerator from data.js to util.js, delete console.log() in main.js
