+++ js/main.js
@@ -1,3 +1,3 @@
 import { getPhotosDataArray, LOADED_PHOTOS_COUNT } from './data.js';
 
-console.log(getPhotosDataArray(LOADED_PHOTOS_COUNT));
+getPhotosDataArray(LOADED_PHOTOS_COUNT);
